feat: implements the basic project structure
