--- authorization-service/src/main/java/com/example/authorization_service/config/AuthorizationServerConfig.java
@@ -1,44 +1,0 @@
-package com.example.authorization_service.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.oauth2.server.authorization.config.annotation.web.configurers.OAuth2AuthorizationServerConfigurer;
-import org.springframework.security.oauth2.server.authorization.settings.ClientSettings;
-import org.springframework.security.oauth2.server.authorization.settings.OAuth2TokenFormat;
-import org.springframework.security.oauth2.server.authorization.settings.TokenSettings;
-import org.springframework.security.web.SecurityFilterChain;
-
-import java.time.Duration;
-
-@Configuration
-@EnableWebSecurity
-public class AuthorizationServerConfig {
-
-    @Bean
-    @Order(1)
-    SecurityFilterChain authorizationSecurityFilterChain(HttpSecurity http) throws Exception{
-
-        OAuth2AuthorizationServerConfigurer authorizationServerConfigurer =
-                OAuth2AuthorizationServerConfigurer.authorizationServer();
-
-        return http.build();
-    }
-
-    @Bean
-    public TokenSettings tokenSettings() {
-        return TokenSettings.builder()
-                            .accessTokenFormat(OAuth2TokenFormat.SELF_CONTAINED)
-                            .accessTokenTimeToLive(Duration.ofMinutes(60))
-                            .build();
-    }
-
-    @Bean
-    public ClientSettings clientSettings() {
-        return ClientSettings.builder()
-                             .requireAuthorizationConsent(false)
-                             .build();
-    }
-}
