+++ src/main/java/com/example/authorization_service/config/AuthorizationServerConfig.java
@@ -1,0 +1,95 @@
+package com.example.authorization_service.config;
+
+import com.example.authorization_service.domain.LoginSocialSuccessHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.core.GrantedAuthorityDefaults;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.oauth2.server.authorization.config.annotation.web.configurers.OAuth2AuthorizationServerConfigurer;
+import org.springframework.security.oauth2.server.authorization.settings.ClientSettings;
+import org.springframework.security.oauth2.server.authorization.settings.OAuth2TokenFormat;
+import org.springframework.security.oauth2.server.authorization.settings.TokenSettings;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
+import org.springframework.security.web.util.matcher.MediaTypeRequestMatcher;
+
+import java.time.Duration;
+
+@Configuration
+@EnableWebSecurity
+public class AuthorizationServerConfig {
+
+    @Bean
+    @Order(1)
+    SecurityFilterChain authorizationSecurityFilterChain(HttpSecurity http, LoginSocialSuccessHandler successHandler) throws Exception{
+
+        var authorizationServerConfigurer = OAuth2AuthorizationServerConfigurer.authorizationServer();
+
+        http
+            .securityMatcher(authorizationServerConfigurer.getEndpointsMatcher())
+            .with(authorizationServerConfigurer, (authorizationServer) ->
+                                                                    authorizationServer.oidc(Customizer.withDefaults())
+            )
+            .authorizeHttpRequests(authorize ->
+                                                                        authorize.anyRequest().authenticated()
+            )
+            .exceptionHandling(exceptions ->
+                                                    exceptions.defaultAuthenticationEntryPointFor(
+                                                            new LoginUrlAuthenticationEntryPoint("/login"),
+                                                            new MediaTypeRequestMatcher(MediaType.TEXT_HTML)));
+
+        return http.build();
+    }
+
+    @Bean
+    SecurityFilterChain securityFilterChain(HttpSecurity http,
+                                            LoginSocialSuccessHandler successHandler) throws Exception{
+        return http
+                .csrf(AbstractHttpConfigurer::disable)
+                .httpBasic(Customizer.withDefaults())
+                .formLogin(configure -> configure.loginPage("/login"))
+                .authorizeHttpRequests(authorize -> {
+                    authorize.requestMatchers(HttpMethod.GET, "/login").permitAll();
+                    authorize.requestMatchers(HttpMethod.POST, "/user/create").permitAll();
+                    authorize.anyRequest().authenticated();
+                })
+                .oauth2Login(oauth2 -> {
+                    oauth2.loginPage("/login").successHandler(successHandler);
+                })
+                .build();
+
+    }
+
+    @Bean
+    public TokenSettings tokenSettings() {
+        return TokenSettings.builder()
+                            .accessTokenFormat(OAuth2TokenFormat.SELF_CONTAINED)
+                            .accessTokenTimeToLive(Duration.ofMinutes(60))
+                            .build();
+    }
+
+    @Bean
+    public ClientSettings clientSettings() {
+        return ClientSettings.builder()
+                             .requireAuthorizationConsent(false)
+                             .build();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder(10);
+    }
+
+    @Bean
+    public GrantedAuthorityDefaults grantedAuthorityDefaults() {
+        return new GrantedAuthorityDefaults("");
+    }
+}
