+++ src/main/java/com/example/authorization_service/dto/ClientDTO.java
