+++ src/main/java/com/example/authorization_service/entity/ClientEntity.java
