+++ src/main/java/com/example/authorization_service/repository/CustomRegisteredClientRepository.java
