+++ src/main/java/com/example/authorization_service/service/ClientService.java
